requirements.txt
APIFlask==2.0.2
Bottleneck==1.3.7
Flask==2.2.2
GDAL==3.6.2
loguru==0.7.2
lru_dict==1.2.0
matplotlib==3.8.0
natsort==7.1.1
numba==0.58.1
numpy==1.24.3
opencv-python==4.9.0.80
pandas==2.2.2
Pillow==10.3.0
PyYAML==6.0.1
scikit-learn==1.3.0
scipy==1.13.0
seaborn==0.13.2
skimage==0.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
tsnecuda==3.0.1
xarray==2023.11.0